logout api integrate

+++ assets/js/check-login-ajax.js
@@ -1,7 +1,13 @@
+var secret = "Goal(lasfk!@#$%^&*()FHKBKahksaf+_)(*&^%$#fdaskfa)Goal";
+var secret2 = "Josh(fafgY/xx$<y3nXJnA,$IYgzf$H@<x_YUxL4yz89r5uxOR%+)Talks";
+
 $(document).ready(() => {
   let userID =
     document.cookie.match &&
     document.cookie.match(new RegExp("loginUserId" + `=([^;]+)`));
+  let authToken =
+    document.cookie.match &&
+    document.cookie.match(new RegExp("AuthToken" + `=([^;]+)`));
   var currentUserId = userID && userID[1];
   if (!currentUserId) {
     alert("user not logged in");
@@ -19,8 +25,23 @@ $(document).ready(() => {
   }
 
   $("#logout").click(function(e) {
-    e.preventDefault();
-    deleteAllCookies();
-    window.location.replace("/");
+    if (currentUserId) {
+      var apiUrl = "https://goal.joshtalks.org/api/app/user/GetUserLogout";
+      logoutData = {
+        ApiKey: SHA1(secret2 + apiUrl.split("/")[apiUrl.split("/").length - 1]),
+        UserId: currentUserId,
+        UserType: 0,
+        AuthToken: authToken[1]
+      };
+
+      $.post(apiUrl, logoutData, function(data, status) {
+        if (data.returncode == 200) {
+          deleteAllCookies();
+          window.location.replace("/");
+        } else {
+          alert(data.msg);
+        }
+      });
+    }
   });
 });
